refactor: move Notifications to ui package
