+++ app/src/main/AndroidManifest.xml
@@ -26,7 +26,7 @@
         <activity android:name=".github.LoginWebViewActivity" />
         <activity android:name=".ProfileActivity" />
         <activity
-              android:name=".NotificationActivity"
+              android:name=".ui.NotificationActivity"
               android:label="@string/notifs"
               android:theme="@style/MaterialDrawerTheme.Light.DarkToolbar"></activity>
           <activity
