+++ app/src/main/java/gitlin/kothub/ui/NotificationActivity.kt
@@ -1,24 +1,17 @@
-package gitlin.kothub
+package gitlin.kothub.ui
 
 import android.support.v7.app.AppCompatActivity
 import android.os.Bundle
-import android.util.Log
-import gitlin.kothub.github.api.data.UserSummary
-import gitlin.kothub.github.api.userSummary
-import kotlinx.android.synthetic.main.activity_notifs.*
-import gitlin.kothub.utilities.*
 
 import org.jetbrains.anko.AnkoLogger
 import org.jetbrains.anko.debug
 import android.view.Gravity
 import android.widget.*
+import gitlin.kothub.R
 import gitlin.kothub.adapters.NotificationAdapter
-import gitlin.kothub.adapters.PinnedRepositoryAdapter
-import gitlin.kothub.github.api.data.Notification
 import gitlin.kothub.github.api.data.Notifications
-import gitlin.kothub.ui.AppDrawer
-import gitlin.kothub.ui.DrawerData
 import kotlinx.android.synthetic.main.toolbar.*
+import kotlinx.android.synthetic.main.activity_notifs.*
 import gitlin.kothub.github.api.*
 
 class NotificationActivity : AppCompatActivity(), AnkoLogger {
